refactor: 사실 Output인데 Input이라고 적은 폴더명 수정
